Simplified the search functions and made a more user friendly one
The original one was purely designed to be used internally, the get
one however is designed for external use. Will return NULL in case
it isn't in the tree. Please don't call this before inserting as the
binary tree will search the tree for the key up front anyway, even
slightly more optimized that is.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>